back-end/repository_system/requirements.txt
asgiref==3.8.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
colorama==0.4.6
cors==1.0.1
crispy-bootstrap5==2024.2
cryptography==44.0.1
distlib==0.3.8
Django==5.0.4
django-cors-headers==4.7.0
django-crispy-forms==2.3
django-deep-serializer==0.1.3
django-environ==0.12.0
django-filter==25.1
django-jazzmin==3.0.1
django-multiselectfield==0.1.13
django-phonenumber-field==8.0.0
djangorestframework==3.15.2
filelock==3.13.4
future==1.0.0
gevent==24.11.1
greenlet==3.1.1
idna==3.10
iniconfig==2.0.0
pyjwt==1.3.1
packaging==24.2
phonenumbers==8.13.47
pillow==10.3.0
platformdirs==4.2.0
pluggy==1.5.0
psycopg==3.2.3
psycopg2-binary==2.9.10
pycparser==2.22
PySocks==1.7.1
pytest==8.3.4
pytest-mock==3.14.0
requests==2.32.3
requests-file==2.1.0
setuptools==75.8.0
sqlparse==0.5.0
tldextract==5.1.3
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.3.0
virtualenv==20.25.3
virtualenvwrapper-win==1.2.7
zope.event==5.0
zope.interface==7.2
